+++ CHANGELOG.md
@@ -14,6 +14,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Refactor integration tests for _SQL Server Reporting Services_ and _Power BI_
     _Report Server_ ([issue #2431](https://github.com/dsccommunity/SqlServerDsc/issues/2431)).
 
+### Added
+
+- SqlServerDsc
+  - Added integration tests for `SqlDatabase` that exercise `PsDscRunAsCredential`
+    and `dsc config` usage with a separate parameters file for secure credentials.
+
 ## [17.4.0] - 2026-01-19
 
 ### Added